requirements.txt
anyio==4.6.2.post1
azure-core==1.31.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
msal==1.31.0
msal-extensions==1.2.0
parse==1.20.2
pip==24.2
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
requests==2.32.3
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
